+++ src/main/java/com/vaadin/example/gxt/companydashboard/client/RadarDataProperties.java
@@ -37,6 +37,8 @@
  */
 package com.vaadin.example.gxt.companydashboard.client;
 
+import java.math.BigDecimal;
+
 import com.google.gwt.editor.client.Editor.Path;
 import com.sencha.gxt.core.client.ValueProvider;
 import com.sencha.gxt.data.shared.ModelKeyProvider;
@@ -47,10 +49,10 @@ import com.sencha.gxt.data.shared.PropertyAccess;
  */
 public interface RadarDataProperties extends PropertyAccess<RadarData> {
 
-  ValueProvider<RadarData, Double> data();
+	ValueProvider<RadarData, BigDecimal> data();
 
-  ValueProvider<RadarData, String> name();
+	ValueProvider<RadarData, String> name();
 
-  @Path("name")
-  ModelKeyProvider<RadarData> nameKey();
+	@Path("name")
+	ModelKeyProvider<RadarData> nameKey();
 }
